typecasting

+++ QSET/ARM_V1/driver/lib/libusb_arm.c
@@ -88,7 +88,7 @@ struct udev_pkt_status *armDev_getStatusPkt(armDev_t *pDev){
 int armDev_setServo(armDev_t *pDev, enum eArmServos servo, uint32_t val_us){
     struct udev_pkt_ctrl ctrl;
     ctrl.hdr.pkt_typ = ePktTypeSrvo;
-    ctrl.hdr.ctrl_typ = servo;
+    ctrl.hdr.ctrl_typ = (uint8_t)servo;
     ctrl.servo_ctrl = val_us;
     return armdev_send(pDev, &ctrl);
 }
@@ -97,7 +97,7 @@ int armDev_setMtr(armDev_t *pDev, enum eArmMotors mtr, struct udev_mtr_ctrl *pVa
     if(pVals == NULL) return -1;
     struct udev_pkt_ctrl ctrl;
     ctrl.hdr.pkt_typ = ePktTypeMtr;
-    ctrl.hdr.ctrl_typ = mtr;
+    ctrl.hdr.ctrl_typ = (uint8_t)mtr;
     memcpy(&ctrl.mtr_ctrl, pVals, sizeof(struct udev_mtr_ctrl));
     memcpy(&pDev->pkt_mtr[mtr], pVals, sizeof(struct udev_mtr_ctrl));
     return armdev_send(pDev, &ctrl);
@@ -118,7 +118,7 @@ int armDev_updateMtr(armDev_t *pDev, enum eArmMotors mtr, float pos, float vel){
 
     struct udev_pkt_ctrl ctrl;
     ctrl.hdr.pkt_typ = ePktTypeMtr;
-    ctrl.hdr.ctrl_typ = mtr;
+    ctrl.hdr.ctrl_typ = (uint8_t)mtr;
     memcpy(&ctrl.mtr_ctrl, &pDev->pkt_mtr[mtr], sizeof(struct udev_mtr_ctrl));
     return armdev_send(pDev, &ctrl);
 }
